docshell/test/mochitest/file_bug1450164.html

<html>
  <head>
    <script>
     function go() {
       var a = window.history.state;
       window.history.replaceState(a, "", "1");
       var ok = opener.ok;
       var SimpleTest = opener.SimpleTest;
       ok("Addition of history in unload did not crash browser");
       SimpleTest.finish();
     }
    </script>
  </head>
  <body onunload="go()">
  </body>
</html>
